+++ lib/player.rb
@@ -1,0 +1,10 @@
+class Player
+  def get_player_input
+    letter = gets.chomp
+    letter[0]
+  end
+
+  def get_name
+    gets.chomp
+  end
+end
